fix: correct dtype casts for autocast modes
The layers assumed autocast was active. Casts are explicit now, so a run
produces the same result with and without `torch.autocast`.

GitLab-MR: !61
GitOrigin-RevId: df350ad1ec867aecc8b6a3d0fe846b951c022813

+++ tensorrt_bionemo/models/boltz1/modeling.py
@@ -404,7 +404,7 @@ class Boltz1(nn.Module):
                          None),
             x_pred=struct_module_output["sample_atom_coords"],
             feature_dict=feed_dict,
-            pred_distogram_logits=pair_distogram,
+            pred_distogram_logits=pair_distogram.float(),
             multiplicity=diffusion_samples,
             attn_metadata=attn_metadata,
             all_reduce_params=all_reduce_params,
